mousePress and noLoop demos now work the same way as the processing demos, the size given to canvas() now correctly sets the size of the canvas rather than the jframe which included the title bar

+++ demos/noLoop.py
@@ -4,7 +4,7 @@ sys.path.append(os.path.pardir + "/lib")
 from graphicsWrapper import *
 
 canvas(640, 360)
-y = height() * 0.5  # Size should be the first statement
+y = int(height() * 0.5)  # Size should be the first statement
 stroke(255)          # Set line drawing color to white
 noLoop()
